Change input file loading, fix bugs in xrefs_2_inchikey function
Input files are loaded from a directory, and csv headers are skipped. The function xrefs_2_inchikey is modified to return a list of all the unique _ids of the query hits, instead of a single _id value. (Previous function was incorrectly returning only the result of the last executed query.) Now, if a document has multiple _id hits, duplicate documents are yielded.